requirements.txt
google-auth==2.23.2
google-cloud==2.3.3
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
redis==3.5.2
